Modularize dom elements and Book class
