+++ modules/displayTime.js
@@ -1,0 +1,6 @@
+import { DateTime } from 'luxon';
+// show date
+elem.dateElem.innerText = new Date();
+setInterval(() => {
+  elem.dateElem.innerText = new Date();
+}, 1000);
